Fix handling of REM, for formatting and validation
Issue#51  https://github.com/mvextensions/mvbasic/issues/51

REM is not handled correctly both from a formatting perspective,
as well as when parsing code for labels.  This patch treats REM
as a comment when it is at the start of a line.  Could also consider
adding support for CRT "HELLO" ; REM THIS IS A COMMENT

^\\s*REM\\s.* is a better regex to max against REM as a comment:

^      Start of line
\\s*  <0 or more whitespace characters>
REM literal REM or rem
\\s*  <0 or more whitespace characters (including \n)>
.*     Whatever remains on the rest of the line