Update server.ts
I found 1 additional bug with VS Code and the data index.  It is missing from the documentation, but it is actually required for it to work.  The bud is that without data set: label,insertText, and documentation work, but detail does not.  If you have multiple items, it uses detail from the last array position for all items.  I also found that the numbers don't do anything for the order of the results returned.  My solution is to use `data: i+1` so that it is handled on the fly.  I know that my VS Code updated a few days ago, so not sure if this is a new bug, because I didn't notice this behavior during my initial testing last week.  I also noticed that the `kind` of icon do not change anymore.