Update server.ts
Changed code to do a count on Intellisense array before the push to assign a IntellisensePosition var in the push that will increment by 1.  This avoids having to keep track of positions in different parts of the code or trying to statically assign them like 999 trying to stay out of range.  This also avoids the user keeping this correct in their files.  This method should probably also be used in the language and custom words sections.