v2.0.11 release (#125)
* Update server.ts

I found 1 additional bug with VS Code and the data index.  It is missing from the documentation, but it is actually required for it to work.  The bud is that without data set: label,insertText, and documentation work, but detail does not.  If you have multiple items, it uses detail from the last array position for all items.  I also found that the numbers don't do anything for the order of the results returned.  My solution is to use `data: i+1` so that it is handled on the fly.  I know that my VS Code updated a few days ago, so not sure if this is a new bug, because I didn't notice this behavior during my initial testing last week.  I also noticed that the `kind` of icon do not change anymore.

* Update server.ts

Changed code to do a count on Intellisense array before the push to assign a IntellisensePosition var in the push that will increment by 1.  This avoids having to keep track of positions in different parts of the code or trying to statically assign them like 999 trying to stay out of range.  This also avoids the user keeping this correct in their files.  This method should probably also be used in the language and custom words sections.

* Update server.ts

switched to single line `data: Intellisense.length + 1,` in all 3 sections that touch Intellisense (language, customer words, and custom functions)

* Update jBASELanguage.json

removed all occurrences of `index` since it is no longer required.

* Shipping v2.0.11

Co-authored-by: Brent Blair <bblair@encompass.com>